+++ scripts/verifyContent.py
@@ -16,7 +16,7 @@ for dir in dirs:
       missing_changelog_entries.add(resource_name.decode())
   
   has_import_file = os.path.isfile(f"{dir}/import.sh")
-  if not has_import_file:
+  if not has_import_file and resource_name.decode() != "pingfederate_keypairs_signing_csr":
     no_import_dirs.add(dir)
 
 if len(missing_changelog_entries) > 0: