+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.13.0 (Unreleased)
 ### Resources
+* **New Resource:** `pingfederate_keypairs_signing_csr` ([#306]([https](https://github.com/pingidentity/terraform-provider-pingfederate/pull/306)))
 * **New Resource:** `pingfederate_sp_target_url_mappings` ([#273]([https](https://github.com/pingidentity/terraform-provider-pingfederate/pull/273)))
 * **New Resource:** `pingfederate_keypairs_ssl_server_settings` ([#272]([https](https://github.com/pingidentity/terraform-provider-pingfederate/pull/272)))
 * **New Resource:** `pingfederate_oauth_authentication_policy_contract_mapping` ([#262]([https](https://github.com/pingidentity/terraform-provider-pingfederate/pull/262)))